% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_update_role.R
\name{add_fourier_vars}
\alias{add_fourier_vars}
\title{computes variables suitable for seasonality predictors}
\usage{
add_fourier_vars(data_to_use = data1, vc = var_controls)
}
\arguments{
\item{data_to_use}{defaults to data1; suitable for MMM and with variables matching the vc table}

\item{vc}{deafults to var_controls; tibble with varname and role indicating which
variable is the time_id column}
}
\description{
computes variables suitable for seasonality predictors
}
