% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_update_role.R
\name{add_groups_and_sort}
\alias{add_groups_and_sort}
\title{sorts and groups data per the variable controls table}
\usage{
add_groups_and_sort(data_to_use = data1, vc = var_controls)
}
\arguments{
\item{data_to_use}{defaults to data1;  a tibble or tidytable suitable for MMM and matching the
variables listed in vc (the variable congtrol table)}

\item{vc}{defaults to var_controls; a tibble with varname, role, role2, describing
the roles variables will play in the MMM}
}
\value{
a tibble
}
\description{
sorts and groups data per the variable controls table
}
