% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_update_role.R
\name{adjust_workflow_for_groups_not_in_data}
\alias{adjust_workflow_for_groups_not_in_data}
\title{checks random terms in workflow_control against data; if fewer than 2 levels for the groups, removes the random terms from search}
\usage{
adjust_workflow_for_groups_not_in_data(
  workflow_controls = workflow_controls,
  data_of_interest = data1
)
}
\arguments{
\item{workflow_controls}{defaults to workflow_controls; a tibble with Value and R_name describing
the various model search options}

\item{data_of_interest}{defaults to data1;  a tibble or tidytable suitable for MMM and matching the
variables listed in vc (the variable congtrol table)}
}
\value{
a tibble
}
\description{
checks random terms in workflow_control against data; if fewer than 2 levels for the groups, removes the random terms from search
}
