% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{assemble_workflow}
\alias{assemble_workflow}
\title{creates a tidymodels workflow from a a recipe and a model}
\usage{
assemble_workflow(this_formula = built_formula, recipe_to_use = recipe3)
}
\arguments{
\item{this_formula}{defaults to built_formula; a string suitable to be coerced to a formula for lmer}

\item{recipe_to_use}{defaults to recipe3; a recipe with roles for predictor, outcome, time_id, and groups}
}
\description{
creates a tidymodels workflow from a a recipe and a model
}
