% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{asymptote}
\alias{asymptote}
\title{dial (ie from the tune package) for asymptote}
\usage{
asymptote(range = c(100, 300))
}
\arguments{
\item{range}{defaults to c(100,300), should be a length2 numeric vector that has the minimum and maximum possible asymptotes for a given variable transformation.  In general the asymptote should be higher than the max value in the historical data.}
}
\value{
a quant param dial
}
\description{
dial (ie from the tune package) for asymptote
}
