% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{bake.step_adstock}
\alias{bake.step_adstock}
\title{bake method for step_adstock}
\usage{
\method{bake}{step_adstock}(object, new_data, ...)
}
\arguments{
\item{object}{is recipe}

\item{new_data}{data.frame or tibble to be baked}
}
\value{
data that is transformed
}
\description{
bake method for step_adstock
}
