% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{bake.step_saturation}
\alias{bake.step_saturation}
\title{The saturation bake method}
\usage{
\method{bake}{step_saturation}(object, new_data, ...)
}
\arguments{
\item{object}{recipe or step}

\item{mdata}{the data fed to the recipe or step}
}
\value{
a transformed dataset
}
\description{
The saturation bake method
}
