% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_update_role.R
\name{bulk_add_role}
\alias{bulk_add_role}
\title{loops over the variable control table to assign \emph{second} rolls within a recipe. These roles control transformations applied and construction of model formulas}
\usage{
bulk_add_role(
  this_recipe,
  vars_to_append_roles = var_controls$varname[!is.na(var_controls$role2)],
  roles_to_be_appended = var_controls$role2[!is.na(var_controls$role2)]
)
}
\arguments{
\item{this_recipe}{a recipe object}

\item{vars_to_append_roles}{Defaults to var_controls$varname\link{!is.na(var_controls$role2)}; a vector of variable names that will have updated roles}

\item{roles_to_be_appended}{Defaults to var_controls$role2\link{!is.na(var_controls$role2)}; a vector of roles that match the vars_to_append_roles vector}
}
\value{
a recipe
}
\description{
loops over the variable control table to assign \emph{second} rolls within a recipe. These roles control transformations applied and construction of model formulas
}
