% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_update_role.R
\name{check_if_needs_tune}
\alias{check_if_needs_tune}
\title{check to see if a recipe needs to be finalized before baking}
\usage{
check_if_needs_tune(recipe_to_check)
}
\arguments{
\item{recipe_to_check}{a recipe}
}
\value{
booelan true or false
}
\description{
check to see if a recipe needs to be finalized before baking
}
\details{
This function wraps prep(recipe_to_check) in tryCatch to produce a TRUE/FALSE on
if the recipe still has tune() in some of its parameters.
This will be inefficient in the situation where the data in the recipe is large or the
prep steps are lengthy
}
