% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{create_dials_from_wf_and_controls}
\alias{create_dials_from_wf_and_controls}
\title{Sets hyperparameter ranges for tune grid}
\usage{
create_dials_from_wf_and_controls(
  workflow = recipe3,
  control_ranges = transform_controls
)
}
\arguments{
\item{workflow}{model workflow ore recipe, defaults to recipe3}

\item{control_ranges}{defaults to transform_controls; a df with ranges by role2 of variable}
}
\description{
Sets hyperparameter ranges for tune grid
}
