% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_update_role.R
\name{create_recipe}
\alias{create_recipe}
\title{reads the variable controls configuration and dataset to create a recipe,
which may include tuning dials}
\usage{
create_recipe(data_to_use = data1, vc = var_controls, mc = transform_controls)
}
\arguments{
\item{data_to_use}{defaults to data1, should be a tibble or tidytable that
has variable names and structure useable as template for the MMM data}

\item{vc}{defaults to var_controls, should be a tibble describing individual
variable roles (ie role and role 2 and varname are included inthe tibble).}

\item{mc}{defaults to transform_controls; a tibble with hyperparameter ranges
by role (ie media type)}
}
\value{
a recipe with roles, adstock and saturation steps defined by controls table
}
\description{
reads the variable controls configuration and dataset to create a recipe,
which may include tuning dials
}
