% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{get_adstock}
\alias{get_adstock}
\title{function used to compute adstock transformation}
\usage{
get_adstock(x, retain)
}
\arguments{
\item{x}{numeric vector or time series of values to be adstocked}

\item{retain}{between 0 and 1, the amount of previous time points value that effects the current time point}
}
\value{
a numeric vector that shows the adstocked values of x
}
\description{
function used to compute adstock transformation
}
