% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{get_control}
\alias{get_control}
\title{Pulls values from the control tibble by name; convenience function for working with the control tibble.}
\usage{
get_control(this_control, control = workflow_controls)
}
\arguments{
\item{this_control}{is the string containing the control of interest}

\item{control}{defaults to workflow_controls; is a tibble containing columns R_name and Value,}
}
\value{
The value of control$Value when control$R_name==this_control; typically will be a character, frequently requires coercion to other types.
}
\description{
Pulls values from the control tibble by name; convenience function for working with the control tibble.
}
