% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{get_outcome_vector}
\alias{get_outcome_vector}
\title{pulls outcome variable from recipe using role== group or role==time_id}
\usage{
get_outcome_vector(base_recipe = recipe3)
}
\arguments{
\item{base_recipe}{defaults to recipe3; a recipe produced in the mostlytidyMMM format}
}
\value{
a character vector
}
\description{
pulls outcome variable from recipe using role== group or role==time_id
}
