% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{get_predictors_vector}
\alias{get_predictors_vector}
\title{pulls predictor variable names from a recipe except those that have role2 \%in\% c('group','time')}
\usage{
get_predictors_vector(base_recipe = recipe3)
}
\arguments{
\item{base_recipe}{defaults to recipe3; is the recipe to pull predictor names from}
}
\value{
a vector of variablnames that have role == 'predictor'
}
\description{
pulls predictor variable names from a recipe except those that have role2 \%in\% c('group','time')
}
