% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{get_train_saturation}
\alias{get_train_saturation}
\title{a monotonic saturation function with monotonic first deriviate used in recipes for preprocssing MMM data}
\usage{
get_train_saturation(x, asymptote, saturation_speed)
}
\arguments{
\item{x}{vector of numeric value to saturate}

\item{asymptote}{the asymptote of the transformed values; should probably exceed the max value in the modeled data}

\item{saturation_speed}{higher values approach the asymptote faster}
}
\value{
a numeric vector of transformed values
}
\description{
a monotonic saturation function with monotonic first deriviate used in recipes for preprocssing MMM data
}
