% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{make_bound_statements}
\alias{make_bound_statements}
\title{converts sign constraints in the variable configuration table to rethinking::ulam ready constraints.}
\usage{
make_bound_statements(variable_controls = var_controls)
}
\arguments{
\item{variable_controls}{Defaults to var_controls; a tibble containing start_name and varname, and sign, which is a column with constraints as '>=0' or '+'. Currently only supports >=0 or <=0, with +,- being used as alias for those.}
}
\value{
a list of boundary statements suitable for rethinking::ulam
}
\description{
converts sign constraints in the variable configuration table to rethinking::ulam ready constraints.
}
