% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{make_list_of_fft_formulae}
\alias{make_list_of_fft_formulae}
\title{create a list of model formula (per the rules for lmer()) with different
seasonality specifications as listed in the configuration tibble.}
\usage{
make_list_of_fft_formulae(vc = workflow_controls, recipe_to_use = recipe3)
}
\arguments{
\item{vc}{defaults to workflow_controls; a tibble with R_name and Value columns
where R_name contains the control name and Value is it's value}

\item{recipe_to_use}{defaults to recipe3; a recipes::recipe with pre-modeling transformations and
variable roles; the outcome, predictor, group, and time_id roles are used in determining
the formula.}
}
\value{
a list with names formulae and configs, formulae is a list of strings suitable
to be coerced to a formula and passed to lmer.  configs is a list of tibbles with information like vc,
but reset to match a single seasonality specification.  If search_seasonality is set to FALSE in
the config table, the two lists have a single formula containing the specified seasonality
and no random terms.  A note is printed to the console in this case.
}
\description{
create a list of model formula (per the rules for lmer()) with different
seasonality specifications as listed in the configuration tibble.
}
