% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{make_list_of_rands_formula}
\alias{make_list_of_rands_formula}
\title{produce a list of  formula for testing combinations of random effects}
\usage{
make_list_of_rands_formula(
  seasonality_formula = best_seas_formula,
  vc = best_seas_vc
)
}
\arguments{
\item{seasonality_formula}{defaults to best_seas_formula; a string containing the formula for
all fixed effects in the model plus the seasonaity terms (which may include random slopes).
Typically will be produced by tuning a workflowset to identify the best seasonality specification.}

\item{vc}{defaults to best_seas vc; a tibble with a single seasonality specification typically determined
by the seasonality_search process}
}
\value{
a list of strings suitable to be coerced to formula for lmer
and the specifications for random effects.  In the case where the control
search_randoms = 'FALSE' , the string in seasonality_formula is returned in a list of 1
}
\description{
produce a list of  formula for testing combinations of random effects
}
