% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{make_prior_statements}
\alias{make_prior_statements}
\title{Creates sampling statements for user defined priors on individual variables.}
\usage{
make_prior_statements(variable_controls = var_controls)
}
\arguments{
\item{variable_controls}{Defaults to var_controls; a tibble containing start_name and varname, prior, and prior_sd, where the column prior is the mean of a normal dist.}
}
\value{
a list of sampling statements suitable for rehtinking::ulam()
}
\description{
Creates sampling statements for user defined priors on individual variables.
}
