% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bound_statements.R
\name{predict.ulam}
\alias{predict.ulam}
\title{a prediction function for ulam objects, as comes from rethinking::ulam}
\usage{
\method{predict}{ulam}(ulamobj, new_data, n = 1000, reduce = T, conf = 0.95)
}
\arguments{
\item{ulamobj}{the out of rethinking::ulam}

\item{new_data}{the dataset on which to draw samples and make predictions}

\item{n}{the number of samples to draw}

\item{reduce}{defuaults to TRUE; a boolean that indicates if the samples should be returned or a mean and high credibility interval}

\item{conf}{defaults to .95, indicates the desired converage of credibility interval when reduce=T}
}
\value{
if reduce=T a tibble with pred, pred_lower_conf, pred_upper_conf columns
else if reduce =F the output of link (a matrix with sampled predictions)
}
\description{
a prediction function for ulam objects, as comes from rethinking::ulam
}
