% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{prep.step_adstock}
\alias{prep.step_adstock}
\title{Prep method for step_adstock}
\usage{
\method{prep}{step_adstock}(x, training, info = NULL, ...)
}
\arguments{
\item{x}{step or recipe}

\item{training}{dataset used for prep}
}
\value{
a recipe step of type adstock
}
\description{
Prep method for step_adstock
}
