% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{prep.step_saturation}
\alias{prep.step_saturation}
\title{The saturation step prep method}
\usage{
\method{prep}{step_saturation}(x, training, info = NULL, ...)
}
\arguments{
\item{x}{an object (step or recipe)}

\item{training}{tibble used to provide data for training}
}
\value{
a prepped recipe or step
}
\description{
The saturation step prep method
}
