% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{print.step_adstock}
\alias{print.step_adstock}
\title{custom print method for adstock steps}
\usage{
\method{print}{step_adstock}(x, width = max(20, options()$width - 35), ...)
}
\arguments{
\item{x}{is recipe or step}
}
\value{
descriptive output to console
}
\description{
custom print method for adstock steps
}
