% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{print.step_saturation}
\alias{print.step_saturation}
\title{print method for saturation steps}
\usage{
\method{print}{step_saturation}(x, width = max(20, options()$width - 35), ...)
}
\arguments{
\item{x}{step}
}
\value{
string
}
\description{
print method for saturation steps
}
