% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_update_role.R
\name{rename_columns_per_controls}
\alias{rename_columns_per_controls}
\title{renames columns in data per the controls file}
\usage{
rename_columns_per_controls(working_df, variable_controls = var_controls)
}
\arguments{
\item{working_df}{is the dataset that needs to be renamed in a tibble or data.frame}

\item{variable_controls}{Defaults to var_controls; a tibble containing start_name and varname, with varname being the desired final name of each columns}
}
\value{
a tibble with (potentially) renamed columns
}
\description{
renames columns in data per the controls file
}
