% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{retention}
\alias{retention}
\title{establishing retention as a dial using new_quant_param}
\usage{
retention(range = c(0, 0.8))
}
\arguments{
\item{range}{is the allowed range of retention for a specific hyperparameter}
}
\value{
a quantitative tunable hyperparameter
}
\description{
establishing retention as a dial using new_quant_param
}
