% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{step_adstock_new}
\alias{step_adstock_new}
\title{internal function used in create recipe steps}
\usage{
step_adstock_new(
  terms,
  role,
  trained,
  retention,
  adstocks,
  groups,
  options,
  skip,
  id
)
}
\value{
a recipe sep of type saturation
}
\description{
internal function used in create recipe steps
}
