% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{step_saturation_new}
\alias{step_saturation_new}
\title{internal function used in create recipe steps}
\usage{
step_saturation_new(
  terms,
  role,
  trained,
  asymptote,
  saturation_speed,
  hills,
  options,
  skip,
  id
)
}
\value{
a recipe sep of type saturation
}
\description{
internal function used in create recipe steps
}
