% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{tunable.step_adstock}
\alias{tunable.step_adstock}
\title{tunable method for step_adstock}
\usage{
\method{tunable}{step_adstock}(x, ...)
}
\arguments{
\item{x}{is a step or recipe}
}
\value{
a tibble interpretable by dials and tune as hyperparameter information
}
\description{
tunable method for step_adstock
}
