% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_saturation.R
\name{tunable.step_saturation}
\alias{tunable.step_saturation}
\title{tunable method for saturation steps -- used by tune package to identify what each hyperparameter is.}
\usage{
\method{tunable}{step_saturation}(x, ...)
}
\arguments{
\item{x}{step or recipe}
}
\value{
tibble that tune functions can evaluate as hyperparameter information
}
\description{
tunable method for saturation steps -- used by tune package to identify what each hyperparameter is.
}
